Backend/requirements.txt
fastapi==0.115.12                    
uvicorn[standard]
uvicorn==0.34.3                         
python-dotenv==1.1.0                             
pandas==2.3.0                       
sqlmodel==0.0.24                      
motor==3.3.1                    
pymongo[srv]==4.5.0        
beanie==1.25.0                 
pydantic>=2.6,<3.0   
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
cryptography==45.0.4  
python-multipart-0.0.20